Prepared exp function but still with decimal part problem

+++ formula_translation/formula_translation/exponential.h
@@ -1,1 +1,1 @@
-double exponential(int x);
+float exponential(int x);
